Feito funcionalidades principais básicas e primeira questão do desafio
