+++ src/Mission.ts
@@ -1,5 +1,5 @@
 import moment from 'moment'
-import { Student } from './Students'
+import { Student } from './students'
 import { Teacher } from './Teacher'
 
 moment.locale("pt-br")
