+++ students.json
@@ -78,5 +78,105 @@
       "dançar",
       "comer"
     ]
+  },
+  {
+    "id": "1",
+    "name": "Kamila",
+    "email": "kamila@gmail.com",
+    "birthDate": "23/04/1988",
+    "hobbies": [
+      "Jogar",
+      "Assistir filmes/séries"
+    ]
+  },
+  {
+    "id": "2",
+    "name": "Lais",
+    "email": "lais@gmail.com",
+    "birthDate": "10/10/1995",
+    "hobbies": [
+      "dançar",
+      "comer"
+    ]
+  },
+  {
+    "id": "1",
+    "name": "Kamila",
+    "email": "kamila@gmail.com",
+    "birthDate": "23/04/1988",
+    "hobbies": [
+      "Jogar",
+      "Assistir filmes/séries"
+    ]
+  },
+  {
+    "id": "2",
+    "name": "Lais",
+    "email": "lais@gmail.com",
+    "birthDate": "10/10/1995",
+    "hobbies": [
+      "dançar",
+      "comer"
+    ]
+  },
+  {
+    "id": "1",
+    "name": "Kamila",
+    "email": "kamila@gmail.com",
+    "birthDate": "23/04/1988",
+    "hobbies": [
+      "Jogar",
+      "Assistir filmes/séries"
+    ]
+  },
+  {
+    "id": "2",
+    "name": "Lais",
+    "email": "lais@gmail.com",
+    "birthDate": "10/10/1995",
+    "hobbies": [
+      "dançar",
+      "comer"
+    ]
+  },
+  {
+    "id": "1",
+    "name": "Kamila",
+    "email": "kamila@gmail.com",
+    "birthDate": "23/04/1988",
+    "hobbies": [
+      "Jogar",
+      "Assistir filmes/séries"
+    ]
+  },
+  {
+    "id": "2",
+    "name": "Lais",
+    "email": "lais@gmail.com",
+    "birthDate": "10/10/1995",
+    "hobbies": [
+      "dançar",
+      "comer"
+    ]
+  },
+  {
+    "id": "1",
+    "name": "Kamila",
+    "email": "kamila@gmail.com",
+    "birthDate": "23/04/1988",
+    "hobbies": [
+      "Jogar",
+      "Assistir filmes/séries"
+    ]
+  },
+  {
+    "id": "2",
+    "name": "Lais",
+    "email": "lais@gmail.com",
+    "birthDate": "10/10/1995",
+    "hobbies": [
+      "dançar",
+      "comer"
+    ]
   }
 ]
